% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBrAPIConnection}
\alias{getBrAPIConnection}
\title{Get Known BrAPI Connection (by database name)}
\usage{
getBrAPIConnection(name = NULL)
}
\arguments{
\item{name}{The name of the known BrAPI server}
}
\value{
BrAPIConnection
}
\description{
Get a BrAPI Connection for a pre-defined database by name.  Use the
\code{listBrAPIConnections()} function to get a list of all pre-defined
databases.  This function will return the connection, with all of
its properties set, for the named database.
}
\examples{
wheat <- getBrAPIConnection("T3/Wheat")
cassava <- getBrAPIConnection("Cassavabase")
grin <- getBrAPIConnection("USDA-GRIN Wheat")

}
