% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBrAPIConnections}
\alias{getBrAPIConnections}
\title{Get all Known BrAPI Connections}
\usage{
getBrAPIConnections()
}
\value{
list(BrAPIConnection)
}
\description{
Return a list of all known BrAPI Connections
}
\examples{
connections <- getBrAPIConnections()

}
