% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_connections.R
\name{KNOWN_BRAPI_CONNECTIONS}
\alias{KNOWN_BRAPI_CONNECTIONS}
\title{Known BrAPI Connections
A list of pre-configured BrAPIConnections to known BrAPI servers}
\usage{
KNOWN_BRAPI_CONNECTIONS
}
\description{
These are used in the getBrapiConnection() function when a user requests a connection by name
}
\keyword{internal}
