% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{BrAPIConnection}
\alias{BrAPIConnection}
\title{BrAPI Connection Class}
\description{
A Reference Class representing a BrAPI connection
}
\details{
All fields are required. The properties defined in the connection will be used
to construct the final URL to a BrAPI endpoint: $protocol://$host/$path/$version/{...call}
}
\section{Fields}{

\describe{
\item{\code{protocol}}{The HTTP protocol - either "http" or "https"}

\item{\code{host}}{The hostname of the server to connect to}

\item{\code{path}}{The base path to the BrAPI endpoints (not including the version)}

\item{\code{version}}{The version of BrAPI to use}
}}


