% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{createBrAPIConnection}
\alias{createBrAPIConnection}
\title{Create BrAPI Connection}
\usage{
createBrAPIConnection(
  host = NULL,
  protocol = NULL,
  path = NULL,
  version = NULL
)
}
\arguments{
\item{host}{The BrAPI server hostname}

\item{protocol}{(optional) the HTTP protocol (either 'http' or 'https') (DEFAULT: https)}

\item{path}{(optional) the base path to the BrAPI endpoints (without the version) (DEFAULT: /brapi/)}

\item{version}{(optional) the BrAPI version to use (DEFAULT: v2)}
}
\value{
BrAPIConnection
}
\description{
Create a BrAPI connection to the specified host
}
\examples{
wheat <- createBrAPIConnection("wheat.triticeaetoolbox.org")
oatv1 <- createBrAPIConnection("oat.triticeaetoolbox.org", version="v1")

}
