% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{getBrAPIConnection}
\alias{getBrAPIConnection}
\title{Get Known BrAPI Connection}
\usage{
getBrAPIConnection(name = NULL)
}
\arguments{
\item{name}{The name of the known BrAPI server}
}
\value{
BrAPIConnection
}
\description{
Get the connection details for a known BrAPI server
}
\examples{
wheat <- getBrAPIConnection("T3/Wheat")
cassava <- getBrAPIConnection("Cassavabase")

}
