% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/known_connections.R
\docType{data}
\name{KNOWN_BRAPI_CONNECTIONS}
\alias{KNOWN_BRAPI_CONNECTIONS}
\title{Known BrAPI Connections}
\format{
An object of class \code{list} of length 8.
}
\usage{
KNOWN_BRAPI_CONNECTIONS
}
\description{
A list of pre-configured BrAPIConnections to known BrAPI servers
}
\keyword{datasets}
