% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listBrAPIConnections}
\alias{listBrAPIConnections}
\title{List Known BrAPI Connections}
\usage{
listBrAPIConnections()
}
\description{
List the known BrAPI connections that can be used in \code{getBrAPIConnection(name)}.
}
\examples{
listBrAPIConnections()

}
